C11-159: extract browser.* socket handlers (split into 2 files)
84 cases + 112 methods/helpers relocated. Split across BrowserHandlers.swift
(slice + 50 methods) and BrowserQueryHandlers.swift (62 methods) to stay under
the ~3k LOC per-file ceiling (1938 / 1751). Because the slice and the methods
now span two files, browser handler methods are internal rather than private
(cohesive browser subsystem, module-internal only). Router wired by prefix.
TerminalController.swift: 13222 -> 9675 LOC (below the DX-4 ~10k target).
